requirements.txt
graphene==3.2.2
fastapi==0.95.1
starlette-graphene3==0.6.0
uvicorn==0.21.1
sqlalchemy==2.0.11
psycopg[binary,pool]==3.1.13  # to avoid dependency on libpq-dev, not compiling psycopg3 from source
argon2-cffi==21.3.0
PyJWT==2.7.0
python-dotenv==1.0.0